+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added the salesChannel field to Organization type
+
 ## [0.23.1] - 2022-10-07
 
 ### Fixed